extraction-service/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
playwright==1.42.0
